API

Documentation for VarianceComponentModels.jl's types and methods.

Index

VarianceComponentModels.TwoVarCompModelRotate

VarianceComponentModels.TwoVarCompVariateRotate

VarianceComponentModels.VarianceComponentModel

VarianceComponentModels.VarianceComponentVariate

VarianceComponentModels.fit_mle!

VarianceComponentModels.fit_reml!

VarianceComponentModels.mle_fs!

VarianceComponentModels.mle_mm!

Types

VarianceComponentModels.VarianceComponentModel — Type

VarianceComponentModel stores the model parameters of a variance component model.

Fields

B: p x d mean parameters

Σ: tuple of d x d variance component parameters

A: constraint matrix for vec(B)

sense: vector of characters '=', '<' or '>'

b: constraint vector for vec(B)

lb: lower bounds for vec(B)

ub: upper bounds for vec(B)

VarianceComponentModels.VarianceComponentVariate — Type

VarianceComponentVariate stores the data of a variance component model.

Feilds

Y: n x d responses

X: n x p predictors

V: tuple of n x n covariance matrices

VarianceComponentModels.TwoVarCompModelRotate — Type

TwoVarCompModelRotate stores the rotated two variance component model.

Fields

Brot: rotated mean parameters B * eigvec

eigval: eigenvalues of eig(Σ[1], Σ[2])

eigvec: eigenvectors of eig(Σ[1], Σ[2])

logdetΣ2: log-determinant of Σ[2]

VarianceComponentModels.TwoVarCompVariateRotate — Type

TwoVarCompVariateRotate stores the rotated two variance component data.

Fields

Yrot: rotated responses eigvec * Y

Xrot: rotated covariates eigvec * X

eigval: eigenvalues of eig(V[1], V[2])

eigvec: eigenvectors of eig(V[1], V[2])

logdetV2: log-determinant of V[2]

Functions

VarianceComponentModels.mle_fs! — Function

mle_fs!(vcmodel, vcdatarot; maxiter, solver, qpsolver, verbose)

Find MLE by Fisher scoring algorithm.

Input

vcmodel: two variane component model VarianceComponentModel, with

vcmodel.B and vcmodel.Σ used as starting point

vcdatarot: rotated two varianec component data TwoVarCompVariateRotate

Keyword

maxiter::Int: maximum number of iterations, default is 1000

solver::Symbol: backend nonlinear programming solver, :Ipopt (default) or :Knitro

qpsolver::Symbol: backend quadratic programming solver, :Ipopt (default) or :Gurobi or Mosek

verbose::Bool: display information

Output

maxlogl: log-likelihood at solution

vcmodel: VarianceComponentModel with updated model parameters

Σse=(Σse[1],Σse[2]): standard errors of estimate Σ=(Σ[1],Σ[2])

Σcov: covariance matrix of estimate Σ=(Σ[1],Σ[2])

Bse: standard errors of estimate B

Bcov: covariance of estimate B

VarianceComponentModels.mle_mm! — Function

mle_mm!(vcmodel, vcdatarot; maxiter, qpsolver, verbose)

Find MLE by minorization-maximization (MM) algorithm.

Input

vcmodel: two variane component model VarianceComponentModel, with

vcmodel.B and vcmodel.Σ used as starting point

vcdatarot: rotated two varianec component data TwoVarCompVariateRotate

Keyword

maxiter::Int: maximum number of iterations, default is 1000

qpsolver::Symbol: backend quadratic programming solver, :Ipopt (default) or :Gurobi or Mosek

verbose::Bool: display information

Output

maxlogl: log-likelihood at solution

vcmodel: VarianceComponentModel with updated model parameters

Σse=(Σse[1],Σse[2]): standard errors of estimate Σ=(Σ[1],Σ[2])

Σcov: covariance matrix of estimate Σ=(Σ[1],Σ[2])

Bse: standard errors of estimate B

Bcov: covariance of estimate B

Reference

H. Zhou, L. Hu, J. Zhou, and K. Lange (2015) MM algorithms for variance components models. http://arxiv.org/abs/1509.07426

VarianceComponentModels.fit_mle! — Function

fit_mle!(vcmodel, vcdata; algo)

Find MLE of variane component model.

Input

vcmodel: two variane component model VarianceComponentModel, with

vcmodel.B and vcmodel.Σ used as starting point

vcdata: two varianec component data VarianceComponentVariate

Keyword

algo::Symbol: algorithm, :FS (Fisher scoring) for :MM

(minorization-maximization algorithm)

Output

maxlogl: log-likelihood at solution

vcmodel: VarianceComponentModel with updated model parameters

Σse=(Σse[1],Σse[2]): standard errors of estimate Σ=(Σ[1],Σ[2])

Σcov: covariance matrix of estimate Σ=(Σ[1],Σ[2])

Bse: standard errors of estimate B

Bcov: covariance of estimate B

VarianceComponentModels.fit_reml! — Function

fit_reml!(vcmodel, vcdata; algo)

Find restricted MLE (REML) of variane component model.

Input

vcmodel: two variane component model VarianceComponentModel, with

vcmodel.B and vcmodel.Σ used as starting point

vcdata: two varianec component data VarianceComponentVariate

Keyword

algo::Symbol: algorithm, :FS (Fisher scoring) for :MM

(minorization-maximization algorithm)

Output

maxlogl: log-likelihood at solution

vcmodel: VarianceComponentModel with updated model parameters

Σse=(Σse[1],Σse[2]): standard errors of estimate Σ=(Σ[1],Σ[2])

Σcov: covariance matrix of estimate Σ=(Σ[1],Σ[2])

Bse: standard errors of estimate B

Bcov: covariance of estimate B